adding/cleaning up tests
