--- test/fixtures/example1.raml
@@ -1,47 +1,0 @@
-#%RAML 1.0
-title: Example API
-baseUri: http://example.com
-/users:
-  get:
-/json:
-  post:
-    body:
-      application/json:
-        schema: |
-          {
-            "$schema": "http://json-schema.org/draft-04/schema#",
-            "type": "object",
-            "properties": {
-              "hello": {
-                "type": "string"
-              }
-            }
-          }
-/urlencoded:
-  post:
-    body:
-      application/x-www-form-urlencoded:
-        properties:
-          hello:
-            type: string
-            pattern: "[a-z]+"
-/formdata:
-  post:
-    body:
-      multipart/form-data:
-        properties:
-          hello:
-            type: string
-            pattern: "[a-z]+"
-/query:
-  get:
-    queryParameters:
-      hello:
-        type: string
-        pattern: "[a-z]+"
-
-/users/{userId}:
-  uriParameters:
-    userId:
-      type: number
-  get:
